+++ internal/workerproto/README.md
@@ -59,6 +59,8 @@ This message type, sent from the worker, contains an error report intended for t
 ~{"type": "error-report", "kind": "critical", "title": "a serious error", "description": "bad stuff has happened!", "extra": {"number of failures": 1}}
 ```
 
+See [here](https://docs.taskcluster.net/docs/reference/core/worker-manager/api#reportWorkerError) for documentation on worker-manager's `reportWorkerError` API.
+
 ### shutdown
 
 When worker-runner receives this message, and the provider supports it, it will invoke the function `RemoveWorker` from
