+++ tools/worker-runner/errorreport/errorreport.go
@@ -5,9 +5,9 @@ import (
 	"log"
 	"reflect"
 
+	"github.com/pkg/errors"
 	"github.com/taskcluster/taskcluster/v30/clients/client-go/tcworkermanager"
 	"github.com/taskcluster/taskcluster/v30/internal/workerproto"
-	"github.com/taskcluster/taskcluster/v30/tools/worker-runner/provider/provider"
 	"github.com/taskcluster/taskcluster/v30/tools/worker-runner/run"
 	"github.com/taskcluster/taskcluster/v30/tools/worker-runner/tc"
 )
@@ -49,8 +49,21 @@ func HandleMessage(msg workerproto.Message, factory tc.WorkerManagerClientFactor
 		Extra:       extraMsg,
 		Title:       msg.Properties["title"].(string),
 	}
-	err = provider.ReportWorkerError(state, factory, &errorReport)
+	err = ReportWorkerError(state, factory, &errorReport)
 	if err != nil {
 		log.Printf("Error reporting worker error: %v\n", err)
 	}
+}
+
+// ReportWorkerError will send a worker error report to worker-manager
+func ReportWorkerError(state *run.State, factory tc.WorkerManagerClientFactory, payload *tcworkermanager.WorkerErrorReport) error {
+	wc, err := factory(state.RootURL, &state.Credentials)
+	if err != nil {
+		return errors.Wrap(err, "error instanciating worker-manager client")
+	}
+	if _, err = wc.ReportWorkerError(state.WorkerPoolID, payload); err != nil {
+		log.Printf("Error reporting worker error: %v", err)
+		log.Printf("Error payload: %v", payload)
+	}
+	return err
 }
